+++ wacker/src/runtime/host.rs
@@ -1,0 +1,29 @@
+use wasmtime::component::ResourceTable;
+use wasmtime_wasi::{WasiCtx, WasiView};
+use wasmtime_wasi_http::{WasiHttpCtx, WasiHttpView};
+
+pub struct Host {
+    pub table: ResourceTable,
+    pub ctx: WasiCtx,
+    pub http: WasiHttpCtx,
+}
+
+impl WasiView for Host {
+    fn table(&mut self) -> &mut ResourceTable {
+        &mut self.table
+    }
+
+    fn ctx(&mut self) -> &mut WasiCtx {
+        &mut self.ctx
+    }
+}
+
+impl WasiHttpView for Host {
+    fn ctx(&mut self) -> &mut WasiHttpCtx {
+        &mut self.http
+    }
+
+    fn table(&mut self) -> &mut ResourceTable {
+        &mut self.table
+    }
+}
